--- Hygrometer/Common/APIKey/AddressSearchAPI.swift
@@ -1,10 +1,0 @@
-//
-//  AddressSearchAPI.swift
-//  Hygrometer
-//
-//  Created by 김민지 on 2022/07/12.
-//
-
-struct AddressSearchAPI {
-  static let key = "FC4AC640-39EA-35CC-8824-21200BA7D6FE"
-}
