+++ Hygrometer/Common/Model/WeatherRequestModel.swift
@@ -9,6 +9,7 @@ import Foundation
 import CoreLocation
 
 struct WeatherRequestModel: Codable {
-    let lat: Double
-    let lon: Double
+  let lat: Double
+  let lon: Double
+
 }
