+++ Hygrometer/Home/HomeCollectionViewCell.swift
@@ -9,52 +9,52 @@ import UIKit
 import SnapKit
 
 class HomeCollectionViewCell: UICollectionViewCell {
-    // MARK: - Properties
-    static let id = "HomeCollectionViewCell"
+  // MARK: - Properties
+  static let id = "HomeCollectionViewCell"
 
-    lazy var background: UIView = {
-        let view = UIView()
-        view.backgroundColor = .comfortable
-        view.layer.cornerRadius = 23.0
+  lazy var background: UIView = {
+    let view = UIView()
+    view.backgroundColor = .comfortable
+    view.layer.cornerRadius = 23.0
 
-        return view
-    }()
+    return view
+  }()
 
-    lazy var locationLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = .white
-        label.font = .systemFont(ofSize: 18.0, weight: .semibold)
-        label.text = "경기도 수원시 장안구 파장동"
+  lazy var locationLabel: UILabel = {
+    let label = UILabel()
+    label.textColor = .white
+    label.font = .systemFont(ofSize: 18.0, weight: .semibold)
+    label.text = "경기도 수원시 장안구 파장동"
 
-        return label
-    }()
+    return label
+  }()
 
-    // MARK: - Helpers
-    func configureUI(humidity: Int) {
-        contentView.backgroundColor = .white
+  // MARK: - Helpers
+  func configureUI(humidity: Int) {
+    contentView.backgroundColor = .white
 
-        if humidity >= 80 {
-            background.backgroundColor = .veryMosit
-        } else if humidity >= 60 {
-            background.backgroundColor = .moist
-        } else if humidity >= 40 {
-            background.backgroundColor = .comfortable
-        } else if humidity >= 30 {
-            background.backgroundColor = .dry
-        } else {
-            background.backgroundColor = .veryDry
-        }
-
-        contentView.addSubview(background)
-        background.addSubview(locationLabel)
-
-        background.snp.makeConstraints { make in
-            make.edges.equalToSuperview()
-        }
-
-        locationLabel.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
-            make.leading.trailing.equalToSuperview().inset(16)
-        }
+    if humidity >= 80 {
+      background.backgroundColor = .veryMosit
+    } else if humidity >= 60 {
+      background.backgroundColor = .moist
+    } else if humidity >= 40 {
+      background.backgroundColor = .comfortable
+    } else if humidity >= 30 {
+      background.backgroundColor = .dry
+    } else {
+      background.backgroundColor = .veryDry
     }
+
+    contentView.addSubview(background)
+    background.addSubview(locationLabel)
+
+    background.snp.makeConstraints { make in
+      make.edges.equalToSuperview()
+    }
+
+    locationLabel.snp.makeConstraints { make in
+      make.centerY.equalToSuperview()
+      make.leading.trailing.equalToSuperview().inset(16)
+    }
+  }
 }
