+++ SECRET.swift
@@ -7,4 +7,5 @@
 
 import Foundation
 
-let APIKEY: String = "2f94734b59ed393c884ba97fae88e446"
+let weatherAPIKEY: String = "2f94734b59ed393c884ba97fae88e446"
+let searchAPIKEY: String = "FC4AC640-39EA-35CC-8824-21200BA7D6FE"
